requirements.txt
# streamlit
# google-generativeai
# openai
# weaviate-client
# tqdm
# python-dotenv
# anthropic
# pdfminer.six
# pi_heif
# unstructured_inference
# unstructured-pytesseract
# unstructured
# # onnx
# # onnxruntime
# tesseract
# poppler-utils
# onnxruntime-tools
# pdfplumber
# Pillow
# nltk

streamlit
google-generativeai
tesseract
unstructured
Pillow
nltk
pdfminer.six
pi_heif
unstructured_inference
unstructured-pytesseract
pdf2image
python-dotenv